moved GitNotes in pages/notes
