+++ client/src/pages/Notes/GitVersionControl/GitNotes.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useTheme } from '../context/ThemeContext';
+import { useTheme } from "../../../context/ThemeContext";
 
 const gitVersionControl = [
   {
